api [nfc]: Move SystemGroupName.displayName out to widgets/content.dart
This UI logic was the one place in lib/api/model/permission.dart
that used the app's localizations, which depend on Flutter (#2335).
Its one caller is in widgets/content.dart, so move it there
as a private extension.
